Updating endpoints path, adding admin routes for ROLES, USERS

+++ src/db/models/api_token.go
@@ -11,7 +11,7 @@ import (
 type APIToken struct {
 	ID        uint32    `gorm:"primary_key" json:"id"`
 	Token     string    `json:"token,omitempty" gorm:"uniqueIndex"`
-	UserID    uint32    `json:"-"`
+	UserID    int64     `json:"-"`
 	User      *User     `gorm:"foreignkey:UserID" json:"-" validate:"-"`
 	RoleID    uint32    `json:"-"`
 	Role      *Role     `gorm:"foreignkey:RoleID" json:"-" validate:"-"`
